+++ CHANGES.rst
@@ -3,10 +3,12 @@ Changes
 =======
 
 0.0.4:
+ - Added support for DeepTemp, DeepSquare, and ShallowTemp models.
  - Added support for Mazurka models.
  - Added support for exporting data from tempograms.
  - Added support for framewise normalization in tempograms.
  - Moved to TensorFlow 1.14.0.
+ - Print number of model parameters.
 
 0.0.3:
  - Added flag ``--interpolate`` for ``tempo`` to increase accuracy.